% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densify_prune.R
\name{densify_prune}
\alias{densify_prune}
\title{Function to densify and prune the original input matrix}
\usage{
densify_prune(original_data, documentation, optimum = 1)
}
\arguments{
\item{original_data}{A data frame with the glottocodes as row names and variable names as column names.}

\item{documentation}{A data frame containing the documentation generated during the densification process.}

\item{optimum}{The iteration number representing the optimum quality score. The function will densify the matrix and prune languages/variables up to this iteration. Default is set to 1 (output is original data frame).}
}
\value{
A pruned and densified data frame.
}
\description{
This function takes the original data matrix and the documentation generated from the densification process. It densifies the matrix based on the optimum quality score and removes languages and variables that were pruned during the iterative process.
}
\examples{
# Assuming 'original_data' and 'documentation' are prepared
# pruned_densified <- densify_prune(original_data, documentation, optimum = 2)
}
\seealso{
\code{\link{densify_steps}}
}
